Moves system endpoints into the system package
