+++ stack/rest/src/main/java/org/apache/usergrid/rest/system/ApplicationsResource.java
@@ -17,14 +17,14 @@
  *  * directory of this distribution.
  *
  */
-package org.apache.usergrid.rest;
+package org.apache.usergrid.rest.system;
 
 import com.sun.jersey.api.json.JSONWithPadding;
 import org.apache.usergrid.corepersistence.service.StatusService;
-import org.apache.usergrid.persistence.Entity;
 import org.apache.usergrid.persistence.EntityManager;
-import org.apache.usergrid.persistence.core.util.StringUtils;
 import org.apache.usergrid.persistence.model.util.UUIDGenerator;
+import org.apache.usergrid.rest.AbstractContextResource;
+import org.apache.usergrid.rest.ApiResponse;
 import org.apache.usergrid.rest.security.annotations.RequireSystemAccess;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
